pyproject.toml
[build-system]
requires = ["hatchling"]        
build-backend = "hatchling.build"

[project]
name = "lazi"
version = "0.1.0"
description = "Build, read, manage widgets within your projects on Lazi via the Lazi SDK"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Jacey Simpson", email = "jaceysimps@gmail.com"}
]
keywords = ["SDK", "Project Management", "API"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Topic :: Software Development :: Libraries :: Python Modules"
]

dependencies = [
    "fastapi>=0.104.0",
    "pydantic>=2.0.0",
    "motor>=3.3.0",
    "pymongo>=4.5.0",
    "python-dotenv>=1.0.0",
    "certifi>=2023.7.22",
    "boto3>=1.28.0",
    "botocore>=1.31.0",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.8.0",
    "python-multipart>=0.0.6"
]

[project.optional-dependencies]
dev = ["pytest", "black", "ruff", "mypy", "pre-commit"]
docs = ["mkdocs", "mkdocs-material"]

[project.urls]
Repository = "https://github.com/Jacey1225/personal-manager-02"
Documentation = "https://jacey1225.github.io/personal-manager-02/"

[tool.hatch.build.targets.wheel]
packages = ["api/services/lazi_sdk"]